rag_api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.6
requests==2.32.3
pydantic==2.8.2
pyjwt==2.9.0
redis==5.0.7
